+++ LethalConstellations/_Release Files/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.1.7]
+ - Fixed issue with LethalConfig soft compatibility.
+ - Added more compatibility for LethalMoonUnlocks
+
 ## [0.1.6]
  - Publicized some things for future compatibility with LethalMoonUnlocks by explodingMods (xCore)
  - fixed typo in a method name